move 'reduce' tests to separate file
