+++ types/ramda/tsconfig.json
@@ -185,6 +185,7 @@
         "test/propOr-tests.ts",
         "test/props-tests.ts",
         "test/propSatisfies-tests.ts",
-        "test/range-tests.ts"
+        "test/range-tests.ts",
+        "test/reduce-tests.ts"
     ]
 }
